+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- createUserExperimental to resolve only after SDK is INITIALIZED
+
+### Changed
+
+- createUserExperimental does not start the SDK by default
+
 ## [4.7.2] - 2022-03-10
 
 ### Fixed